Improve handling of decimal separator (input and output):
When reading with dec="auto", decimal separator is determined by looking at numbers in header fields.  This is part of the new function extract.dec.and.units(), which replaces extract.units().

When writing, any commas in header fields with "numbers" (still treated as strings) are replaced with period.  (Data already has always been written with period.)